refactor rtttl to pull out player
