+++ examples/example1
@@ -1,10 +1,12 @@
 #!/usr/bin/env lua
 
 local rtttl = require 'rtttl'
+local play = require('sox-player').play
 
 m = "JingleBell:d=8,o=5,b=112:32p,a,a,4a,a,a,4a,a,c6,f.,16g,2a,a#,a#,a#.,16a#,a#,a,a.,16a,a,g,g,a,4g,4c6"
 rt = rtttl.parse(m)
 rtttl.print(rt)
 
 -- uses default player
-rtttl.play(rt)
+play(rt)
+
