+++ examples/example3
@@ -1,6 +1,7 @@
 #!/usr/bin/env lua
 
 local rtttl = require 'rtttl'
+local play = require('sox-player').play
 
 m = "Take On Me:o=5,d=8,b=160,b=160:f#,f#,f#,d,p,b4,p,e,p,e,p,e,g#,g#,a,b,a,a,a,e,p,d,p,f#,p,f#,p,f#,e,e,f#,e,f#,f#,f#,d,p,b4,p,e,p,e,p,e,g#,g#,a,b,a,a,a,e,p,d,p,f#,p,f#,p,f#,e,e5"
 
@@ -8,4 +9,4 @@ rt = rtttl.parse(m)
 rtttl.print(rt)
 
 -- uses default player
-rtttl.play(rt)
+play(rt)
